+++ engine/src/test/java/net/jqwik/engine/properties/arbitraries/DefaultStringArbitraryTests.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.*;
 
 import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.*;
+import net.jqwik.api.constraints.*;
 
 import static net.jqwik.api.ArbitraryTestHelper.*;
 
@@ -145,6 +146,14 @@ class DefaultStringArbitraryTests {
 		for (char c : DefaultCharacterArbitrary.WHITESPACE_CHARS) {
 			assertAtLeastOneGenerated(generator, s -> s.contains(Character.toString(c)));
 		}
+	}
+
+	@Property
+	void randomStringsShouldContainZeroChar(@ForAll @StringLength(min = 1, max = 20) String aString) {
+		Statistics.label("contains 0").collect(aString.contains("\u0000"));
+		Statistics.label("0 at last position").collect(aString.charAt(aString.length() - 1) == '\u0000');
+		Statistics.coverageOf("contains 0", coverage -> coverage.check(true).count(c -> c > 10));
+		Statistics.coverageOf("0 at last position", coverage -> coverage.check(true).count(c -> c > 10));
 	}
 
 }